ADDED: progressBar for timeLeft and using faster timer to poll every 0.1 seconds to update UI

+++ MusicTechQuiz/GameController.h
@@ -17,7 +17,7 @@
 -(void)gameControllerDelegateDidEndWithTotalScore:(NSNumber*)totalScore;
 -(void)gameControllerDelegateTimerDidRunOut;
 @optional
--(void)gameControllerDelegateTimeUpdate:(NSInteger)timeLeft;
+-(void)gameControllerDelegateTimeUpdate:(float)timeLeft;
 @end
 
 @interface GameController : NSObject
